requirements.txt
appdirs==1.4.4
boto==2.49.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Deprecated==1.2.12
distlib==0.3.2
docopt==0.6.2
filelock==3.0.12
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
MouseInfo==0.1.3
mysql-connector-python==8.0.25
Pillow==8.3.0
pipenv==2021.5.29
pipreqs==0.4.10
protobuf==3.17.3
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
PyGithub==1.55
PyJWT==2.1.0
PyMsgBox==1.0.9
PyNaCl==1.4.0
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PyTweening==1.0.3
requests==2.25.1
six==1.16.0
urllib3==1.26.5
virtualenv==20.4.7
virtualenv-clone==0.5.4
Werkzeug==2.0.1
wrapt==1.12.1
yarg==0.1.9
